requirements.txt
websockets==12.0
python-dotenv==1.0.0
supabase==2.3.0
asyncio
json
base64
os
